% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzyclusteringsimilarity.R
\name{andc}
\alias{andc}
\title{Compute Adjusted Normalized Degree Of Concordance}
\usage{
andc(clustering1, clustering2, model, one_sided = FALSE, massage = TRUE)
}
\arguments{
\item{clustering1}{matrix, clustering (possibly fuzzy)}

\item{clustering2}{matrix, clustering (possibly fuzzy)}

\item{model}{string, adjustment model}

\item{one_sided}{logical, use one sided adjustment, defaults to FALSE}

\item{massage}{logical, massage matrix into Julia format, defaults to TRUE
(cols are points, rows are clusters)}
}
\description{
Compute the andc between clustering1 and clustering2 using model for
expectation adjustment.
}
