% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzyclusteringsimilarity.R
\name{endc}
\alias{endc}
\title{Compute Expected Normalized Degree Of Concordance}
\usage{
endc(clustering1, clustering2, model, one_sided = FALSE, massage = TRUE)
}
\arguments{
\item{clustering1}{matrix, clustering (possibly fuzzy)}

\item{clustering2}{matrix, clustering (possibly fuzzy)}

\item{model}{string, adjustment model}

\item{one_sided}{logical, use one sided adjustment, defaults to FALSE}

\item{massage}{logical, massage matrix into Julia format, defaults to TRUE
(cols are points, rows are clusters)}
}
\description{
Compute the endc between random clusters generated
by model. If one sided = TRUE, endc between randomly
generated clusterings and clustering2.
}
