% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzyclusteringsimilarity.R
\name{fuzzy_setup}
\alias{fuzzy_setup}
\title{Setup fuzzyclusteringsimilarity}
\usage{
fuzzy_setup(pkg_check = TRUE, ...)
}
\arguments{
\item{pkg_check}{logical, check for FuzzyClusteringSimilarity.jl
and install if necessary}

\item{...}{Parameters are passed down to JuliaCall::julia_setup}
}
\description{
This function initializes Julia and the FuzzyClusteringSimilarity.jl package.
The first time will be long since it includes precompilation.
Additionally, this will install Julia and the required packages
if they are missing.
}
