% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzyclusteringsimilarity.R
\name{ndc}
\alias{ndc}
\title{ComputeS Normalized Degree Of Concordance}
\usage{
ndc(clustering1, clustering2, massage = TRUE)
}
\arguments{
\item{clustering1}{matrix, clustering (possibly fuzzy)}

\item{clustering2}{matrix, clustering (possibly fuzzy)}

\item{massage}{logical, massage matrix into Julia format, defaults to TRUE
(cols are points, rows are clusters)}
}
\description{
Compute the ndc between clustering1 and clustering2.
}
